refactoring a bit
* Moving listener to their own packages
* fixing bug after having upgraded the simpleslackapi dep.

runtime tested, seems working, but still need to recode the odoo
today/week calls, as I want to get rid of the previous client API I was
using before.
