+++ src/main/groovy/fr/spironet/slackbot/listeners/KibanaListener.groovy
@@ -1,18 +1,19 @@
-package fr.spironet.slackbot
+package fr.spironet.slackbot.listeners
 
 import com.ullink.slack.simpleslackapi.SlackChannel
 import com.ullink.slack.simpleslackapi.SlackPreparedMessage
 import com.ullink.slack.simpleslackapi.SlackSession
 import com.ullink.slack.simpleslackapi.SlackUser
 import com.ullink.slack.simpleslackapi.events.SlackMessagePosted
+import com.ullink.slack.simpleslackapi.listeners.SlackMessagePostedListener
 import fr.spironet.slackbot.webbrowser.WebBrowser
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
-class GrafanaListener implements com.ullink.slack.simpleslackapi.listeners.SlackMessagePostedListener  {
-    private final static Logger logger = LoggerFactory.getLogger(GrafanaListener.class)
+class KibanaListener implements SlackMessagePostedListener  {
+    private final static Logger logger = LoggerFactory.getLogger(KibanaListener.class)
 
-    private def usage = "Usage: !grafana monitoring: returns a screenshot of the current monitoring screen.\n"
+    private def usage = "Usage: !kibana weather: returns a screenshot of the kibana dashboard.\n"
     private def error = "I am not capable of browsing the web currently.\n${this.usage}"
 
 
@@ -26,18 +27,21 @@ class GrafanaListener implements com.ullink.slack.simpleslackapi.listeners.Slack
             return
         }
 
-        if (messageContent.contains("!grafana")) {
+        if (messageContent.contains("!kibana")) {
             try {
-                def match = messageContent =~ /\!grafana (\S+)/
+                def match = messageContent =~ /\!kibana (\S+)/
                 def issueKey = match[0][1]
-                // Getting the current state of the monitoring screen
-                if (issueKey == "monitoring") {
+                // Getting the "météo des plateformes"
+                if (issueKey == "weather") {
                     def wb = new WebBrowser()
-                    def screenshot = wb.visitGrafanaMonitoringDashboard()
+                    def screenshot = wb.visitKibanaDashboard()
                     if (screenshot == null) {
                         throw new NullPointerException()
                     }
-                    session.sendFile(channelOnWhichMessageWasPosted, screenshot, "monitoring dashboard")
+                    session.sendFile(
+                            channelOnWhichMessageWasPosted,
+                            screenshot,
+                            "Weather report on our geOrchestra platforms on the past 7 days")
                     return
                 } else {
                     session.sendMessage(channelOnWhichMessageWasPosted,
